feat: 任务 人员

+++ src/controllers/authController.js
@@ -10,16 +10,64 @@ const generateToken = (id) => {
 };
 
 
+const exitUser = async(req, res) => {
+  const username = req.headers['x-username']
+  const currentDate = getCurrentDate()
+  try {
+    const user = await User.findOneAndUpdate(
+      { username: username },
+      {
+        isOnline: false,
+        exitAt: currentDate
+      },
+      { new: true, runValidators: true }
+    );
+
+    if (!user) {
+      return res.status(404).json(apiResponse({ code: 404 }));
+    }
+
+    res.json(apiResponse());
+  } catch (error) {
+    res.status(400).json(apiResponse({ code: 400, message: error.message }));
+  }
+}
+
+const logoutUser = async(req, res) => {
+  const username = req.headers['x-username']
+  const currentDate = getCurrentDate()
+  try {
+    const user = await User.findOneAndUpdate(
+      { username: username },
+      {
+        isLogout: true,
+        logoutAt: currentDate,
+        isOnline: false,
+        exitAt: currentDate
+      },
+      { new: true, runValidators: true }
+    );
+
+    if (!user) {
+      return res.status(404).json(apiResponse({ code: 404 }));
+    }
+
+    res.json(apiResponse());
+  } catch (error) {
+    res.status(400).json(apiResponse({ code: 400, message: error.message }));
+  }
+}
+
 const updateUser = async (req, res) => {
   const username = req.headers['x-username']
-  const id = req.headers['x-id']
   const currentDate = getCurrentDate()
   // 密码 是否在线 注销、登录、退出、时间
   try {
     const product = await User.findOneAndUpdate(
-      { username: username, _id: id },
+      { username: username },
       {
         ...req.body,
+        updatedAt: currentDate,
         exitAt: currentDate
       },
       { new: true, runValidators: true }
@@ -35,17 +83,50 @@ const updateUser = async (req, res) => {
   }
 };
 
+// 获取所有
+const getUsers = async (req, res) => {
+  try {
+     const { keyword } = req.query;
+      const { username, ...filter } = req.body ||  {};
+    
+    // 构建查询条件
+    let query = { ...filter };
+    
+    // 模糊搜索：标题或描述包含关键词
+    if (keyword) {
+      query.$or = [
+        { username: { $regex: keyword, $options: 'i' } },        // 标题模糊匹配
+        // { description: { $regex: taskName, $options: 'i' } }   // 描述模糊匹配
+      ];
+    }
+
+    const users = await User.find({ ...query });
+    // : tasks.filter(item => !item.isDelete)
+    res.json(apiResponse({ users }));
+  } catch (error) {
+    res.status(500).json(apiResponse({ code: 500, message: error.message }));
+  }
+};
+
 
 const deleteUser = async (req, res) => {
   try {
-    const user = await User.findOneAndDelete({
-      _id: req.params.id,
-      userId: req.user._id
-    });
+    const user = await User.find(
+      { username: username },
+    );
 
     if (!user) {
       return res.status(404).json(apiResponse({ code: 404 }));
     }
+  
+    if(!user.isLogout) {
+      return res.status(404).json(apiResponse({ code: 500, message: '只能删除已注销的账号！' }));
+    }
+  
+    await User.findOneAndDelete({
+      username: username
+    });
+
 
     res.json({ message: 'user deleted successfully' });
   } catch (error) {
@@ -76,22 +157,29 @@ const login = async (req, res, next) => {
 
 
     const user = await User.findOne({ username });
-    console.log('cuilanxin user', user, user.isOnline)
+
+    
     if (!user) {
-      return res.status(401).json(apiResponse({
-        code: 401,
+      return res.status(500).json(apiResponse({
+        code: 500,
         message: '账号不存在'
       }));
     }
 
     const isPasswordMatch = await user.comparePassword(password);
     if (!isPasswordMatch) {
-      return res.status(401).json(apiResponse({
-        code: 401,
+      return res.status(500).json(apiResponse({
+        code: 500,
         message: '密码不正确'
       }));
     }
 
+    if(user.isLogout) {
+      return res.status(500).json(apiResponse({
+        code: 500,
+        message: '账号已注销'
+      }));
+    }
     // if (!forceLogin && user.isOnline) {
     //   return res.status(401).json(apiResponse({
     //     code: 401,
@@ -133,7 +221,7 @@ const register = async (req, res, next) => {
       }));
     }
 
-    const { username, password } = req.body;
+    const { username, password, name } = req.body;
 
     // 验证请求数据
     if (!username) {
@@ -159,12 +247,12 @@ const register = async (req, res, next) => {
     }
 
     const currentDate = getCurrentDate()
-
     const user = await User.create({
+      name,
       username,
       /*email,*/
       password,
-      registerAt: currentDate,
+      registerAt: currentDate.split(' ')[0],
       updatedAt: currentDate,
     });
 
@@ -180,4 +268,4 @@ const register = async (req, res, next) => {
 };
 
 
-module.exports = { register, login, updateUser, deleteUser };
+module.exports = { logoutUser, register, login, updateUser, deleteUser, getUsers, exitUser };
